requirements.txt
dlib==19.24.0
face-recognition==1.3.0
numpy
Pillow==9.4.0
open-cv
